+++ src/main/java/com/project/project/repository/PropositionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropositionRepository extends JpaRepository<Proposition, Long> {
+public interface PropositionRepository extends JpaRepository<Proposition, Integer> {
 }
 */
