+++ src/main/java/com/project/project/repository/RequestRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RequestRepository extends JpaRepository <Request, Long> {
+public interface RequestRepository extends JpaRepository <Request, Integer> {
 }
 */
