+++ src/main/java/com/project/project/repository/ResponseRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResponseRepository extends JpaRepository <Response, Long> {
+public interface ResponseRepository extends JpaRepository <Response, Integer> {
 }
 
 */
