+++ src/main/java/com/project/project/repository/SurveyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 
     @Repository
-    public interface SurveyRepository extends JpaRepository<Survey, Long>{
+    public interface SurveyRepository extends JpaRepository<Survey, Integer>{
 }
 
 */
